0.2.0

+++ cbot.py
@@ -5,7 +5,7 @@ Application Name: CaliBot
 Functionality Purpose: An agile Discord Bot to fit Cali's needs
 Version: 0.2.0
 '''
-#3/17/20
+#3/24/20
 
 import discord
 import asyncio
@@ -31,12 +31,8 @@ bot.run('token')
 '''
 
 #Organize code
-#Learn to edit messages to prevent clutter
 #Have `/verify` prompt to specify college
-#Verify #verify works
 
-#Put more detail into `/help`
-#Create a no-reply gmail account
 #Prevent email from spam
 
 class Unbuffered(object):
@@ -145,16 +141,19 @@ async def on_message(message):
     
     if message.content.startswith('/verify'): # Verify command
         ucid = message.content.strip("/verify ")
-        email_addr = f"{ucid}@njit.edu"
-        vCode = send_email(email_addr); ID = message.author.id
-        with open("verify.txt", 'a') as f:
-            f.write(f"{vCode} {email_addr} {message.author.id}\n")
-        temp = await message.channel.send(f"**We've sent a verification code to your email at** ___{email_addr}___**, please copy & paste it below.**")
-        try:
-            await asyncio.wait_for(check_verify(vCode, message, temp), timeout=900) # Purge messages when record is removed from 'verify.txt' otherwise purge in 15 minutes
-        except asyncio.TimeoutError:
-            await message.delete(); await temp.delete()
-        edit_file("verify.txt", f"{vCode} {email_addr} {ID}")
+        if not re.search(r"^\w{0,4}\d{0,4}$", ucid):
+            await message.channel.send("**Invalid command! Please make sure you're typing everything correctly.**", delete_after=30)
+        else:
+            email_addr = f"{ucid}@njit.edu"
+            vCode = send_email(email_addr); ID = message.author.id
+            with open("verify.txt", 'a') as f:
+                f.write(f"{vCode} {email_addr} {message.author.id}\n")
+            temp = await message.channel.send(f"**We've sent a verification code to your email at** ___{email_addr}___**, please copy & paste it below.**")
+            try:
+                await asyncio.wait_for(check_verify(vCode, message, temp), timeout=900) # Purge messages when record is removed from 'verify.txt' otherwise purge in 15 minutes
+            except asyncio.TimeoutError:
+                await message.delete(); await temp.delete()
+            edit_file("verify.txt", f"{vCode} {email_addr} {ID}")
 
     if message.channel.id == 688625250832744449: # Listen for code on NJIT MSA #verify
         eCode = re.search(r"^\d\d\d\d$", message.content)
